proyectos/bucket-tarea/requirements.txt
flask
boto3
python-dotenv
pyjwt
